% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat.aus}
\alias{dat.aus}
\title{Australian state and territory boundaries}
\format{
An \code{sf} object with 11 rows and 2 columns, CRS EPSG:3112
(GDA94 / Geoscience Australia Lambert):
\describe{
\item{name}{State or territory name. Must match \code{name} in \link{dat.timezone.names}.}
\item{geometry}{MULTIPOLYGON boundary.}
}
}
\source{
rnaturalearth package
}
\usage{
dat.aus
}
\description{
Polygons of Australian states, territories and external island territories.
Used by \code{\link[=fire_get_timezone]{fire_get_timezone()}} to find which state a fire is in, so the
matching time zone can be looked up in \link{dat.timezone.names}.
}
\keyword{datasets}
