% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat.fire.polygon}
\alias{dat.fire.polygon}
\title{Example fire boundary}
\format{
\subsection{\verb{An example fire boundary use for testing out some functions}}{

An sf object with 1 rows and 4 columns:
\describe{
\item{FireName}{Name given to the fire}
\item{StartDate}{Fire ignition date}
\item{EndDate}{Fire declared extinguished date}
...
}
}
}
\source{
\url{https://datasets.seed.nsw.gov.au/dataset/fire-history-wildfires-and-prescribed-burns-1e8b6}
}
\usage{
dat.fire.polygon
}
\description{
A fire polygon from NPWS fire history data
}
\keyword{datasets}
