% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat.sentinel2.pathrow}
\alias{dat.sentinel2.pathrow}
\title{Sentinel 2 tiles for Australia}
\format{
\subsection{\verb{Used in a function to find which Sentinel 2 tiles intersect with a fire, then construct paths to search NCI}}{

An sf object with 2155 rows and 3 columns:
\describe{
\item{Name}{Sentinel 2 Tile Name}
...
}
}
}
\source{
\url{https://sentiwiki.copernicus.eu/web/s2-mission}
}
\usage{
dat.sentinel2.pathrow
}
\description{
A subset of Sentinel 2 tiles with Path and Row that intersect Australia
}
\keyword{datasets}
