% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat.timezone.names}
\alias{dat.timezone.names}
\title{Timezone Names for Australian States}
\format{
\subsection{\verb{An example fire boundary use for testing out some functions}}{

\describe{
\item{name}{State Name}
\item{tz_name}{timezone name for posixct tz parameter}
...
}
}
}
\usage{
dat.timezone.names
}
\description{
Timezone Names for Australian States
}
\keyword{datasets}
