% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_copernicus_download.R
\name{fire_copernicus_download}
\alias{fire_copernicus_download}
\title{Download copernicus data}
\usage{
fire_copernicus_download(
  fire_bbox,
  download_df,
  dest_folder,
  user_copernicus,
  password_copernicus
)
}
\arguments{
\item{fire_bbox}{Polygon for cropping images. Same as used for fire_copernicus_search()}

\item{download_df}{Data frame that is the result of fire_copernicus_search()}

\item{dest_folder}{Path where the geotifs will be written}

\item{user_copernicus}{copernicus username}

\item{password_copernicus}{copernicus password}
}
\value{
Writes geotifs to disk
}
\description{
Download data from Copernicus.
Take the ouput of fire_copernicus_search(). Requires username and password to download the data  https://dataspace.copernicus.eu/
}
\examples{
#fire_download_copernicus(fire_bbox = dat.bbox,download_df = dat.sen3.filt,dest_folder = outdir)
}
