% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_download_copernicus.R
\name{fire_download_copernicus}
\alias{fire_download_copernicus}
\title{Download copernicus data}
\usage{
fire_download_copernicus(
  fire_bbox,
  download_df,
  dest_folder,
  user_copernicus = "mstorey@uow.edu.au",
  password_copernicus = "rR$VRQFfFfg2D5"
)
}
\arguments{
\item{fire_bbox}{Polygon for cropping images. Same as used for fire_search_copernicus()}

\item{download_df}{Data frame that is the result of fire_search_copernicus()}

\item{dest_folder}{Path where the geotifs will be written}

\item{user_copernicus}{copernicus username}

\item{password_copernicus}{copernicus password}
}
\value{
Writes geotifs to disk
}
\description{
Download copernicus data
}
\examples{
#fire_download_copernicus(fire_bbox = dat.bbox,download_df = dat.sen3.filt,dest_folder = outdir)
}
