% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_download_stac.R
\name{fire_download_stac}
\alias{fire_download_stac}
\title{Download and save false colour Landsat and Sentinel 2 images.}
\usage{
fire_download_stac(fire_bbox, stac_df, dest_folder)
}
\arguments{
\item{fire_bbox}{Polygon of search area, same polygon used as input to fire_search_stac()}

\item{stac_df}{Data frame that is the output of fire_search_stac}

\item{dest_folder}{the output folder for geotifs.}
}
\value{
Writes geotifs to file
}
\description{
Download and save false colour Landsat and Sentinel 2 images.
}
\examples{
#outdir <- "C:\\temp\\landsat_and_sentinel"
#dir.create(outdir)
#fire_download_stac(dat.bbox,dat.stac,outdir)
}
