% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_download_thredds.R
\name{fire_download_thredds}
\alias{fire_download_thredds}
\title{Download a list of thredds (or other) files}
\usage{
fire_download_thredds(df_download, dest_folder)
}
\arguments{
\item{df_download}{a data frame with a list of path in a path_download column}

\item{dest_folder}{destination folder}
}
\value{
Nothing. Downloads files to destination folder
}
\description{
Download a list of thredds (or other) files
}
\examples{
#x <- fire_process_polygon(dat.fire.polygon,4283,"FireName","StartDate","EndDate",2,2,buffer_km = 20)
#z <- fire_search_landsat(x$fire_bbox)
#fire_download_thredds(z,"C:/data")
}
