% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_geo_png.R
\name{fire_geo_png}
\alias{fire_geo_png}
\title{Title}
\usage{
fire_geo_png(input_path, output_path, poly)
}
\arguments{
\item{input_path}{path to image file written by fire_search_MODIS_VIIRS function}

\item{output_path}{destination path for georeferenced image file}

\item{poly}{bounding box used in fire_search_MODIS_VIIRS function}
}
\value{
.TIF file written to disk
}
\description{
Title
}
\examples{
#
}
