% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_gg_map.R
\name{fire_gg_map}
\alias{fire_gg_map}
\title{Create gg plot with and image and/or polygon}
\usage{
fire_gg_map(image = NULL, prog = NULL, agg = 5)
}
\arguments{
\item{image}{terra::rast() object}

\item{prog}{sf polygons}

\item{agg}{value for terra::aggregate()}
}
\value{
gg plot
}
\description{
Takes a terra raster and plots it using terrainr package. The terra aggregate function is first used to reduce image size for plotting.
If a polygon is also an input, it will be plotted over the image.
}
\examples{
#
}
