% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_GIBS_map.R
\name{fire_GIBS_map}
\alias{fire_GIBS_map}
\title{Create NASA GIBS map with NASA hotspots}
\usage{
fire_GIBS_map(
  fire_bbox,
  date_string,
  wms_lyr_name,
  add_hotspots = F,
  hotspots = NULL
)
}
\arguments{
\item{fire_bbox}{Fire bounding box polygon}

\item{date_string}{Date for wms image request}

\item{wms_lyr_name}{wms layer name. See fire_GIBS_products()}

\item{add_hotspots}{Add VIIRS or MODIS hotspots. TRUE/FALSE}

\item{hotspots}{hotspots sf object from fire_search_hotspots()}
}
\value{
leaflet map
}
\description{
Create NASA GIBS map with NASA hotspots
}
\examples{
# fire_GIBS_map(fire_bbox,date_seq[i],wms,T,hotspots)
}
