% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_GIBS_products.R
\name{fire_GIBS_products}
\alias{fire_GIBS_products}
\title{Print a list of useful NASA GIBS WMS products}
\usage{
fire_GIBS_products()
}
\value{
Print a list
}
\description{
Print a list of useful NASA GIBS WMS products
}
\examples{
fire_GIBS_products()
}
