% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_himawari_search.R
\name{fire_himawari_search}
\alias{fire_himawari_search}
\title{Search for Himiwari data}
\usage{
fire_himawari_search(fire_bbox, start_time, end_time, timestep_minutes = 10)
}
\arguments{
\item{fire_bbox}{An sf polygon to search for images}

\item{start_time}{The first time (posixct) for which to search for images. This will be converted to utc within the function.}

\item{end_time}{The last time (posixct) for which to search for images in local time. This will be converted to utc within the function.}

\item{timestep_minutes}{minutes timestep in multiples of 10. e.g. search for images every 10 minutes or 60 minutes.}
}
\value{
Data frame with thredds path to Himawari data
}
\description{
Search the BOM repository of Himawari 8 and 9 images that are held on NCI THREDDS: https://geonetwork.nci.org.au/geonetwork/srv/eng/catalog.search#/metadata/f8433_0020_1861_5916.
There is a lag of around 2 months for data to appear on this server, so NRT data is also searched for more recent records.
The server will be searched and a data frame of paths will be returned. Each row is path to one different band x time combination.
Images are available every 10 minutes, but put in a bigger timestep to limit the amount of records returned.
}
\examples{
#dat.himiwari <- fire_search_himawari(fire_bbox = dat.bbox,start_time = datestart, end_time = dateend,timestep_minutes = 120)
}
