% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_hotspot_map_v2.R
\name{fire_hotspot_map_v2}
\alias{fire_hotspot_map_v2}
\title{Create geoferenced hotspot and WMS image}
\usage{
fire_hotspot_map_v2(
  fire_bbox,
  start_time,
  end_time,
  mapkey = "a5452249ca7c7a4ee2e1e6da787f57cc",
  dest_folder,
  add_hotspots = T,
  outwidth = 500
)
}
\arguments{
\item{fire_bbox}{Polygon of fire or area to search for images}

\item{mapkey}{A mapkey reuired to download hotspots, https://firms.modaps.eosdis.nasa.gov/api/map_key/}

\item{dest_folder}{Folder to save output tifs other files used in processing, including hotspots}

\item{add_hotspots}{T/F to add or exclude hotspots from output images}

\item{outwidth}{width in pixels of output tif}

\item{start_date}{First date from which to search for images}

\item{end_date}{Final date to search for image}
}
\value{
Write a geotiff to disk. Also writes hotspots csv
}
\description{
Create geoferenced hotspot and WMS image
}
\examples{
#fire_hotspot_map(fire_bbox = dat.bbox,start_date = datestart,end_date = dateend,dest_folder = outdir)
}
