% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_landsat_tif.R
\name{fire_landsat_tif}
\alias{fire_landsat_tif}
\title{Create a rgb false colour image}
\usage{
fire_landsat_tif(
  band1_path,
  band2_path,
  band3_path,
  yaml_path,
  fire_bbox,
  dest_fold
)
}
\arguments{
\item{band1_path}{tif to use for red band}

\item{band2_path}{tif to use for green band}

\item{band3_path}{tif to use for blue band}

\item{yaml_path}{odc-metadata.yaml file}
}
\value{
a 3 band tif
}
\description{
Create a rgb false colour image
}
\examples{
#
}
