% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_read_plot.R
\name{fire_read_plot}
\alias{fire_read_plot}
\title{Read a pre-made fire scan image from database}
\usage{
fire_read_plot(lineid, dbpassword)
}
\arguments{
\item{lineid}{ROS line id}

\item{dbpassword}{password for database}
}
\value{
grob
}
\description{
Read a pre-made fire scan image from database
}
\examples{
#
}
