% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_read_scan.R
\name{fire_read_scan}
\alias{fire_read_scan}
\title{Read a linescan with s3path}
\usage{
fire_read_scan(s3name)
}
\arguments{
\item{s3name}{full https path or tif file name}
}
\value{
terra rast object
}
\description{
Read a linescan with s3path
}
\examples{
#
}
