% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_SAM.R
\name{fire_SAM}
\alias{fire_SAM}
\title{Map fire polygon using Segment Anything}
\usage{
fire_SAM(
  image_path,
  polygons_sf = NULL,
  points_sf = NULL,
  working_dir = tempdir()
)
}
\arguments{
\item{image_path}{Path to tif}

\item{polygons_sf}{sf polygons}

\item{points_sf}{sf points}

\item{working_dir}{directory for temporary files. Default is temp dir}
}
\value{
sf polygon object
}
\description{
Map fire polygon using Segment Anything
}
\examples{
#
}
