% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_save_GIBS_map.R
\name{fire_save_GIBS_map}
\alias{fire_save_GIBS_map}
\title{Capture wms map screenshot and write geotiff}
\usage{
fire_save_GIBS_map(fire_bbox, wms_GIBS_map, dest_folder)
}
\arguments{
\item{fire_bbox}{Fire bounding box polygon}

\item{wms_GIBS_map}{Leaflet map from fire_GIBS_map()}

\item{dest_folder}{Output folder for geotiffs}
}
\value{
write geotiff to disk
}
\description{
Capture wms map screenshot and write geotiff
}
\examples{
# fire_save_GIBS_map(fire_bbox,m,dest_folder)
}
