% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_search_copernicus.R
\name{fire_search_copernicus}
\alias{fire_search_copernicus}
\title{Search copernicus stac browser for sentinel 3 (or other) images}
\usage{
fire_search_copernicus(
  fire_bbox,
  start_date,
  end_date,
  collection_names = c("SENTINEL-3")
)
}
\arguments{
\item{fire_bbox}{Polygon of fire or area to search for images}

\item{start_date}{First date from which to search for images}

\item{end_date}{Final date to search for image}

\item{collection_names}{Defaults to "SENTINEL-3", but can be changed to access other image on copernicus stac. https://documentation.dataspace.copernicus.eu/APIs/STAC.html}
}
\value{
data frame with url to copernicus data
}
\description{
Search copernicus stac browser for sentinel 3 (or other) images
}
\examples{
#dat.sen3 <- fire_search_copernicus(fire_bbox = dat.bbox,start_date = datestart,end_date = dateend)
}
