% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_search_hotpots.R
\name{fire_search_hotspots}
\alias{fire_search_hotspots}
\title{Search NASA FIRMS hotspots API}
\usage{
fire_search_hotspots(
  fire_bbox,
  mapkey,
  start_date,
  end_date,
  dest_fold,
  product_filter = NULL
)
}
\arguments{
\item{fire_bbox}{Bounding sf polygon of search area}

\item{mapkey}{users mapkey requested from https://firms.modaps.eosdis.nasa.gov/api/map_key/}

\item{start_date}{Search start date in YYYY-mm-dd}

\item{end_date}{Search end date in YYYY-mm-dd}

\item{dest_fold}{Folder to temporarily download hotspots files (which are removed in function)}

\item{product_filter}{Regex character string for filter hotspot products, e.g. 'VIIRS', 'MODIS', 'SNPP', 'NOAA20', 'NOAA21'}
}
\value{
sf point object
}
\description{
Search NASA FIRMS hotspots API
}
\examples{
#z <- fire_search_hotspots(x$fire_bbox,mapkey,x$fire_bbox$startdate_search,x$fire_bbox$enddate_search,"C:/temp_data","NRT")
}
