% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_search_landsat.R
\name{fire_search_landsat}
\alias{fire_search_landsat}
\title{Search NCI thredds server for Landsat data}
\usage{
fire_search_landsat(fire_bbox)
}
\arguments{
\item{fire_bbox}{bounding box sf polygon of a fire}
}
\value{
a data frame with download paths of landsat imagery
}
\description{
Search NCI thredds server for Landsat data
}
\examples{
#x <- fire_process_polygon(dat.fire.polygon,4283,"FireName","StartDate","EndDate",2,2,buffer_km = 20)
#z <- fire_search_landsat(x$fire_bbox)
}
