% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_search_progs.R
\name{fire_search_progs}
\alias{fire_search_progs}
\title{Search for fire progression polygons}
\usage{
fire_search_progs(fire_bbox, start_date, end_date)
}
\arguments{
\item{fire_bbox}{A polygon, usually fire bounding box, to search for images}

\item{start_date}{The first date for which to search for images YYYY-mm-dd}

\item{end_date}{The last date for which to search for images YYYY-mm-dd}
}
\value{
sf object
}
\description{
Search for fire progression polygons
}
\examples{
#dat.progs.sf <- fire_search_progs(fire_bbox=dat.bbox,start_date=datestart,end_date=dateend)
}
