% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_search_ros.R
\name{fire_search_ros}
\alias{fire_search_ros}
\title{Search for fire ROS lines by weather conditions}
\usage{
fire_search_ros(
  fire_bbox = wfprogression::fire_bbox_polygon(),
  start_time = as.POSIXct("1950-01-01"),
  end_time = as.POSIXct("2050-01-01"),
  ros = c(0, 100),
  temperature = c(0, 100),
  rh = c(0, 100),
  windspeed = c(0, 100),
  progressions = TRUE,
  dbpassword
)
}
\arguments{
\item{fire_bbox}{A polygon (\code{sf} object). Usually the fire bounding box to search for overlapping fire progression polygons.}

\item{start_time}{The first date/time to search for (\code{POSIXct}, local timezone).}

\item{end_time}{The last date/time to search for (\code{POSIXct}, local timezone).}

\item{ros}{ROS range for search query in km/h}

\item{temperature}{Temperature range for search query in C}

\item{rh}{Relative Humidity range for search query in \%}

\item{windspeed}{Wind speed range for search query in km/h}

\item{progressions}{Logical. If \code{TRUE}, return an \code{sf} object with progressions. If \code{FALSE}, don't return progressions (to save processing time).}

\item{dbpassword}{Database password for connecting to the fire progression database.}
}
\value{
An \code{sf} object (default) or a \code{data.frame} containing fire progression attributes.
}
\description{
This function queries the University of Wollongong fire progression database for ROS lines derived from mapped fire progression.
The database will be searched by weather conditions (ranges), ROS range, location and time.
The result will be a list containing the ROS lines, BARRA weather data and progression polygons

Note: This function requires access to the CERMb fire progression Postgres database via VPN or internal network.
}
\examples{
#
}
