% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_search_scans.R
\name{fire_search_scans}
\alias{fire_search_scans}
\title{Search database for line scans}
\usage{
fire_search_scans(fire_bbox, start_date, end_date)
}
\arguments{
\item{fire_bbox}{A polygon, usually fire bounding box, to search for images}

\item{start_date}{The first date for which to search for images YYYY-mm-dd}

\item{end_date}{The last date for which to search for images YYYY-mm-dd}
}
\value{
sf object
}
\description{
Search database for line scans
}
\examples{
#dat.scans.sf <- fire_search_scans(fire_bbox=dat.bbox,start_date=datestart,end_date=dateend)
}
