% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_search_sentinel2.R
\name{fire_search_sentinel2}
\alias{fire_search_sentinel2}
\title{Search NCI thredds server for sentinel 2 data}
\usage{
fire_search_sentinel2(fire_bbox)
}
\arguments{
\item{fire_bbox}{bounding box sf polygon of a fire}
}
\value{
a data frame with download paths of sentinel 2 imagery
}
\description{
Search NCI thredds server for sentinel 2 data
}
\examples{
#x <- fire_process_polygon(dat.fire.polygon,4283,"FireName","StartDate","EndDate",2,2,buffer_km = 20)
#z <- fire_search_sentinel2(x$fire_bbox)
}
