% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_search_stac.R
\name{fire_search_stac}
\alias{fire_search_stac}
\title{Search Geoscience Aus. STAC for Landsat and Sentinel 2
At the moment this export only 3 bands used to create false colour images}
\usage{
fire_search_stac(
  fire_bbox,
  start_date,
  end_date,
  collection_names = c("ga_ls7e_ard_3", "ga_ls8c_ard_3", "ga_ls9c_ard_3",
    "ga_ls5t_ard_3", "ga_s2am_ard_3", "ga_s2bm_ard_3")
)
}
\arguments{
\item{fire_bbox}{A polygon, usually fire bounding box, to search for images}

\item{start_date}{The first date for which to search for images YYYY-mm-dd}

\item{end_date}{The last date for which to search for images YYYY-mm-dd}

\item{collection_names}{Names of which Landsat and Sentinel 2 products to search for. Defaults to all products from Landsat 5 onwards.}
}
\value{
data frame with image paths
}
\description{
Search Geoscience Aus. STAC for Landsat and Sentinel 2
At the moment this export only 3 bands used to create false colour images
}
\examples{
#dat.stac <- fire_search_stac(fire_bbox = dat.bbox,start_date = datestart,end_date = dateend)
}
