% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_sentinel3_tif.R
\name{fire_sentinel3_tif}
\alias{fire_sentinel3_tif}
\title{Convert a list of Sentinel-3 SEN3.zips to geotiffs}
\usage{
fire_sentinel3_tif(fire_bbox, sen3_folder)
}
\arguments{
\item{fire_bbox}{Bounding box polygon which the output tifs will be cropped to.}

\item{sen3_folder}{path to folder where "SEN3.zips" are stored. All SEN3.zip in folder will be converted.}
}
\value{
Geotiffs written to disk
}
\description{
Convert a list of Sentinel-3 SEN3.zips to geotiffs
}
\examples{
#fire_sentinel3_tif(fire_bbox = dat.fire.polygon,sen3_folder = outdir)
}
