% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_stac_download.R
\name{fire_stac_download}
\alias{fire_stac_download}
\title{Download Landsat and Sentinel-2 images and clip to fire extent.}
\usage{
fire_stac_download(fire_bbox, stac_df, dest_folder)
}
\arguments{
\item{fire_bbox}{An \code{sf} polygon of the fire area of interest. Should match the input used for \code{fire_stac_search()}.}

\item{stac_df}{A data frame produced by \code{fire_stac_search()} containing the band URLs and metadata.}

\item{dest_folder}{Local folder path to save the cropped GeoTIFF images.}
}
\value{
Writes GeoTIFF files to \code{dest_folder}. No object is returned.
}
\description{
Downloads Landsat and Sentinel-2 false-colour images using STAC API output.
Streams selected bands directly from cloud storage using GDAL \verb{/vsicurl/},
crops the imagery to the specified fire bounding box, merges the bands into
a false-colour composite, stretches pixel values to 0-255, and writes the result
as a GeoTIFF to the specified output folder.

This function is intended to be used with the output of \code{fire_stac_search()}.
}
\examples{
\dontrun{
library(sf)
library(dplyr)
# Example bounding box: a square around an area of interest
fire_bbox <- st_as_sfc(st_bbox(c(xmin = 149.1, ymin = -35.4, xmax = 149.2, ymax = -35.3)), crs = 4326) \%>\%
  st_as_sf(crs=4326)

# Define search time window
start_time <- as.POSIXct("2023-01-01", tz = "Australia/Sydney")
end_time <- as.POSIXct("2023-01-07", tz = "Australia/Sydney")

# Run the search
stac_results <- fire_stac_search(fire_bbox, start_time, end_time)

head(stac_results)

outdir <- "C:/temp/landsat_sentinel"
dir.create(outdir)
fire_stac_download(fire_bbox, stac_results, outdir)
}
}
