% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_stac_sample_veg.R
\name{fire_stac_sample_veg}
\alias{fire_stac_sample_veg}
\title{Search for Landsat and Sentinel 2 data}
\usage{
fire_stac_sample_veg(
  sf_object,
  start_time = NULL,
  end_time = NULL,
  collection_name
)
}
\arguments{
\item{sf_object}{An sf object}

\item{start_time}{The first time for which to search for images (posixct required). This will be converted to utc.}

\item{end_time}{The last time for which to search for images (posixct required). This will be converted to utc.}

\item{collection_name}{Names of which Landsat and Sentinel 2 products to search for. Allowed names are: "ga_ls_fc_pc_cyear_3","ga_ls_landcover_class_cyear_3","ga_srtm_dem1sv1_0","ga_ls_fc_3". https://knowledge.dea.ga.gov.au/notebooks/Beginners_guide/03_Products_and_measurements/}
}
\value{
data frame with image paths
}
\description{
Search Geoscience Aus. STAC browser for Landsat and Sentinel 2 images that intersect polygon and match search dates.
This will return a data frame of images path to input to fire_stac_download() function.
A max of 999 records will be returned, so location and or date may need to be limited.
Info on STAC browser is here: https://knowledge.dea.ga.gov.au/guides/setup/gis/stac/
}
\examples{
#dat.stac <- fire_search_stac(fire_bbox = dat.bbox,start_time = datestart,end_time = dateend)
}
