% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_thredds_list.R
\name{fire_thredds_list}
\alias{fire_thredds_list}
\title{List files from a THREDDS catalog URL filtered by a regex pattern}
\usage{
fire_thredds_list(catalog_url, regex)
}
\arguments{
\item{catalog_url}{Character. The URL to the THREDDS catalog HTML page for a specific time (e.g., \verb{https://thredds.nci.org.au/thredds/catalog/ra22/satellite-products/arc/obs/himawari-ahi/fldk/latest/2025/01/01/0000/catalog.html}).}

\item{regex}{Character. A regular expression pattern used to filter the filenames returned. (e.g. \code{ABOM_OBS_} returns observations for each band)}
}
\value{
Character vector of filtered file names from the catalog page, or NA if there was an error accessing the page.
}
\description{
This internal utility function fetches the content of a THREDDS catalog HTML page for Himawari data from BOM,
extracts all file names listed on that page, and returns only those that match a
specified regular expression pattern. It is useful for programmatically listing netcdf files at a path for a specific time
e.g.\verb{https://thredds.nci.org.au/thredds/catalog/ra22/satellite-products/arc/obs/himawari-ahi/fldk/latest/2025/01/01/0000/catalog.html}

The function safely handles errors by returning NA if the catalog cannot be accessed.
It also trims whitespace and filters the filenames according to the regex provided.
}
\examples{
fire_thredds_list(catalog_url="https://thredds.nci.org.au/thredds/catalog/ra22/satellite-products/arc/obs/himawari-ahi/fldk/latest/2025/01/01/0000/catalog.html",
                  regex="ABOM_OBS_")
}
