% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_tidync_safe.R
\name{fire_tidync_safe}
\alias{fire_tidync_safe}
\title{Function to try loading a remote NetCDF file with retry on failure}
\usage{
fire_tidync_safe(url, max_tries = 5, wait_seconds = 10)
}
\arguments{
\item{url}{nc URL, usually on THREDDS}

\item{max_tries}{number of times to try loading nc again in case of error}

\item{wait_seconds}{how long to wait between each try, in seconds}
}
\value{
nc connection
}
\description{
Function to try loading a remote NetCDF file with retry on failure
}
\examples{
#
}
