% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_uv2ds.R
\name{fire_uv2ds}
\alias{fire_uv2ds}
\title{Wind speed and direction from U and V}
\usage{
fire_uv2ds(u, v)
}
\arguments{
\item{u}{u component from BARRA data}

\item{v}{v component from BARRA data}
}
\value{
speed (m/s) and direction
}
\description{
Wind speed and direction from U and V
}
\examples{
# Example U and V components
u <- c(3, -2, 0)
v <- c(4, 2, -5)

# Calculate wind speed and direction
fire_uv2ds(u, v)

# Returns a matrix:
#      dir    speed
# [1,]  36.87 5.00
# [2,] 317.10 2.83
# [3,] 180.00 5.00
}
