% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parked_fire_wms_to_tif_python.R
\name{fire_wms_to_tif_python}
\alias{fire_wms_to_tif_python}
\title{Python based wms to tif}
\usage{
fire_wms_to_tif_python(layers, time, bbox, size, png_file, geotiff_file)
}
\arguments{
\item{layers}{gibs layer name}

\item{time}{date in YYYY-mm-dd/YYYY-mm-dd}

\item{bbox}{bbox coords}

\item{size}{output pixel sizes c(length,width)}

\item{png_file}{output png path}

\item{geotiff_file}{output tif path}
}
\value{
write tif
}
\description{
Python based wms to tif
}
\examples{
#
}
