% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_write_plot.R
\name{fire_write_plot}
\alias{fire_write_plot}
\title{Title}
\usage{
fire_write_plot(tif_path, polygon, fire_bbox, dest_folder)
}
\arguments{
\item{tif_path}{Path to geotiff}

\item{polygon}{Polygon to map, usually fire hsitory boundary}

\item{fire_bbox}{Fire bounding box polygon from fire_bbox_polygon()}

\item{dest_folder}{Path to write jpg plots}
}
\value{
write jpgs to file
}
\description{
Title
}
\examples{
#purrr::map(list.tifs,~fire_write_plot(.x,dat.fire.polygon,dat.bbox,outdir))
}
