% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_barra_sample.R
\name{fn_sample_nc}
\alias{fn_sample_nc}
\title{Function to stream and sample BARRA weather netcdf}
\usage{
fn_sample_nc(nc_conn, datetimeutc, sf_points, varname)
}
\arguments{
\item{nc_conn}{netcdf connection using tidync::tidync()}

\item{datetimeutc}{posixct datetime in utc. Must match https path of nc connection}

\item{sf_points}{sample point (sf object)}

\item{varname}{BARRA variable name. Must match var name used in https path.}
}
\value{
sf object with new columns for sampled variable
}
\description{
Function to stream and sample BARRA weather netcdf
}
\examples{
#
}
