% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uv2ds.R
\name{uv2ds}
\alias{uv2ds}
\title{Wind speed and direction from U and V}
\usage{
uv2ds(u, v)
}
\arguments{
\item{u}{u component from BARRA data}

\item{v}{v component from BARRA data}
}
\value{
speed (m/s) and direction
}
\description{
Wind speed and direction from U and V
}
\examples{
#
}
