% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_barra_sample.R
\name{fire_barra_sample}
\alias{fire_barra_sample}
\title{Function to stream and sample BARRA weather netcdf}
\usage{
fire_barra_sample(
  nc_conn,
  datetimeutc,
  sf_data,
  varname,
  allcells = F,
  timestep,
  extract_fun = "mean"
)
}
\arguments{
\item{nc_conn}{netcdf connection using tidync::tidync()}

\item{datetimeutc}{posixct datetime in utc. Must match https path of nc connection}

\item{sf_data}{sf object with one row with locations to sample barra data}

\item{varname}{BARRA variable name. Must match var name used in https path. e.g. sfcWind (surface wind), tas (temperature), hurs (RH), vas and uas (wind components).}

\item{allcells}{logical. Return all cells in BARRA data that intersect an input polygon? Leave F for point data.}

\item{timestep}{Timestep of data to search for in BARRA. Will accept "hourly", "daily" or "monthly"}

\item{extract_fun}{If not returning all cells for an input polygon, how should sampled cells be summarised in terra::extract, e.g. mean, max}
}
\value{
sf object with new columns for sampled variable
}
\description{
Take an existing nc connection (tidync::tidync()) and sample for the locations in the input sf points
}
\examples{
#
}
