% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_barra_sample_all.R
\name{fire_barra_sample_all}
\alias{fire_barra_sample_all}
\title{Sample an sf object for BARRA data.}
\usage{
fire_barra_sample_all(dat, time_col_utc, barraid = "C2", varnames)
}
\arguments{
\item{dat}{sf object, points lines or polygons}

\item{time_col_utc}{name string of datetime column (posix, utc)}

\item{barraid}{R2 (12 km BARRA product) or C2 (~4km BARRA product)}

\item{varnames}{Vector of barra variable names. BARRA variable name e.g. sfcWind (surface wind), tas (temperature), hurs (RH), vas and uas (wind components).  http://www.bom.gov.au/research/publications/researchreports/BRR-067.pdf}
}
\value{
sample sf object
}
\description{
Sample an sf object for BARRA data.
}
\examples{
#
}
