% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_bbox_polygon.R
\name{fire_bbox_polygon}
\alias{fire_bbox_polygon}
\title{Create a bounding box polygon}
\usage{
fire_bbox_polygon(fire_polygon, bbox_buffer_km, new_crs = NULL)
}
\arguments{
\item{fire_polygon}{An sf object which is the fire boundary polygon (or search area)}

\item{bbox_buffer_km}{How many km to buffer the fire polygon by during the creation of a bounding box.}

\item{new_crs}{New epsg code if transformation is required. Default to 4283 (GDA94)}
}
\value{
A list with one fire polygon and one bounding box polygon
}
\description{
Create a bounding box polygon
}
\examples{
#x <- fire_bbox_polygon(fire_polygon = dat.fire.polygon,bbox_buffer_km = 20,new_crs = 4283)
}
