% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_bbox_polygon.R
\name{fire_bbox_polygon}
\alias{fire_bbox_polygon}
\title{Create a bounding box sf polygon}
\usage{
fire_bbox_polygon(sf_polygon = NULL, bbox_buffer_km = 0, crs = 3112)
}
\arguments{
\item{sf_polygon}{An sf object which is the fire boundary polygon (or search area)}

\item{bbox_buffer_km}{How many km to buffer the fire polygon by during the creation of a bounding box.}

\item{crs}{CRS for buffering (epsg code). Must be projected CRS.}
}
\value{
A bounding box sf polygon with the input polygon's CRS
}
\description{
Takes an sf polygon (e.g. fire boundary), optionally buffers the polygon, create an sf bbox and then convert to sf polygon.
If sf_polygon is empty, a box around Australia is the output
}
\examples{
#fire_bbox_polygon(dat.fire,100,3112)
}
