% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_cermb_weather.R
\name{fire_cermb_weather}
\alias{fire_cermb_weather}
\title{Sample AWS BOM date from CERMB database}
\usage{
fire_cermb_weather(
  sf_point,
  datetime,
  nearest = T,
  buffer_dist_km = 250,
  dbpassword,
  sf_point_id = NULL
)
}
\arguments{
\item{sf_point}{sf point used to find nearest BOM AWS station}

\item{datetime}{vector of posixct. Single or multiple times.}

\item{nearest}{logical. TRUE returns only nearest station. FALSE returns all within buffer.}

\item{buffer_dist_km}{distance in km around sample point to find weather stations. Set to 250 km by default. 250 km is the max distance of a station that this function will find (stations >= 251km aren't sampled).}

\item{dbpassword}{password for the CERMB database}

\item{sf_point_id}{Add point id to output data. Leave as NULL to not add column.}
}
\value{
An \code{sf} object with AWS weather
}
\description{
Sample AWS BOM date from CERMB database
}
\examples{
#
}
\keyword{internal}
