% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_cermb_weather.R
\name{fire_cermb_weather}
\alias{fire_cermb_weather}
\title{Sample AWS BOM date from CERMB database}
\usage{
fire_cermb_weather(sf_point, sf_point_id, datetime, dbpassword)
}
\arguments{
\item{sf_point}{sf point used to find nearest BOM AWS station}

\item{datetime}{vector of posixct}

\item{dbpassword}{password for the CERMB database}
}
\value{
An \code{sf} object with AWS weather
}
\description{
Sample AWS BOM date from CERMB database
}
\examples{
#
}
