% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_get_line_info.R
\name{fire_get_line_info}
\alias{fire_get_line_info}
\title{Get start an end coords and directions of an sf line}
\usage{
fire_get_line_info(sf_lines)
}
\arguments{
\item{sf}{lines input}
}
\value{
data frame
}
\description{
Get start an end coords and directions of an sf line
}
\examples{
#
}
