% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_get_timezone.R
\name{fire_get_timezone}
\alias{fire_get_timezone}
\title{Get Australian time zone of fire_bbox centroid}
\usage{
fire_get_timezone(fire_bbox)
}
\arguments{
\item{fire_bbox}{Fire bounding box polygon}
}
\value{
String of time zone name e.g. "Australia/Sydney"
}
\description{
Get Australian time zone of fire_bbox centroid
}
\examples{
#fire_get_timezone(dat.bbox)
}
