% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_himawari_download.R
\name{fire_himawari_download}
\alias{fire_himawari_download}
\title{Download Himawari images from BOM thredds server}
\usage{
fire_himawari_download(
  fire_bbox,
  df_download,
  bands = c("B07", "B04", "B01"),
  dest_folder
)
}
\arguments{
\item{fire_bbox}{Fire bounding box polygon}

\item{df_download}{Data frame of himawari paths from fire_himiwari_search() function}

\item{bands}{Himawari bands to download. Default is 7, 4 and 1. First band used for night images. Can be a single band.}

\item{dest_folder}{File local destination}
}
\value{
Writes geotifs to disk
}
\description{
Accepts the output of fire_himiwari_search() to clip and download data from BOMS' NCI thredds server
Because Himawari data is course resolution, the fire_bbox is buffered by 100 km before clipping.
Different bands combination can be selected. 7-4-1 is good for seeing active fire.
Output tif will be an RGB for day time images, but single band (first band in the 'bands' parameter) for night time images.
}
\examples{
# fire_download_himiwari(fire_bbox = dat.bbox,df_download = dat.himiwari,dest_folder = outdir)
}
