% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_hotspot_map.R
\name{fire_hotspot_map}
\alias{fire_hotspot_map}
\title{Title}
\usage{
fire_hotspot_map(
  fire_bbox,
  start_date,
  end_date,
  mapkey = "a5452249ca7c7a4ee2e1e6da787f57cc",
  dest_folder
)
}
\arguments{
\item{fire_bbox}{Polygon of fire or area to search for images}

\item{start_date}{First date from which to search for images}

\item{end_date}{Final date to search for image}

\item{mapkey}{A mapkey reuired to download hotspots, https://firms.modaps.eosdis.nasa.gov/api/map_key/}

\item{dest_folder}{Folder to save output tifs other files used in processing, including hotspots}
}
\value{
Write a geotiff to disk. Also writes hotspots csv
}
\description{
Title
}
\examples{
#fire_hotspot_map(fire_bbox = dat.bbox,start_date = datestart,end_date = dateend,dest_folder = outdir)
}
