% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_max_spread_line.R
\name{fire_max_spread_line}
\alias{fire_max_spread_line}
\title{Find maximum fire spread line}
\usage{
fire_max_spread_line(
  polygons,
  time_col,
  id_col,
  convex_hull = T,
  max_only = T,
  internal_only = F,
  min_minutes = 20,
  max_minutes = 300,
  densify_m = 100
)
}
\arguments{
\item{polygons}{Fire polygons sf object}

\item{time_col}{Name of date time column (posix)}

\item{id_col}{Name of unique id column to add to output lines}

\item{convex_hull}{T/F whether to convert fire polygon to convex hull first}

\item{max_only}{T/F return only the max spread line}

\item{internal_only}{T/F return only lines completely within the fire polygon at time 2}

\item{min_minutes}{minimum time interval between progressions}

\item{max_minutes}{maximum time interval between progression}

\item{densify_m}{metres between vertices on polygon edge from which to measure spread}
}
\value{
sf object
}
\description{
Find maximum fire spread line
}
\examples{
#
}
