% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_max_spread_line.R
\name{fire_max_spread_line}
\alias{fire_max_spread_line}
\title{Find line of maximum fire spread}
\usage{
fire_max_spread_line(
  polygons,
  time_col,
  id_col,
  convex_hull = T,
  max_only = T,
  internal_only = F,
  min_minutes = 20,
  max_minutes = 300,
  densify_m = 100
)
}
\arguments{
\item{polygons}{Fire polygons sf object}

\item{time_col}{Name of date time column (posix)}

\item{id_col}{Name of unique id column to add to output lines}

\item{convex_hull}{T/F whether to convert fire polygon to convex hull first}

\item{max_only}{T/F return only the max spread line}

\item{internal_only}{T/F return only lines completely within the fire polygon at time 2}

\item{min_minutes}{minimum time interval between progressions}

\item{max_minutes}{maximum time interval between progression}

\item{densify_m}{metres between vertices on polygon edge from which to measure spread}
}
\value{
sf object
}
\description{
Takes a set of sf polygons with, each with a posixct datetime field, and find the line of maximum difference or spread from one polygon to the next.
The second polygon should completely contain the first, but this may work if that is not the case.
There are some option for calculating this by converting the polygons to convex hulls, which makes the process fast but is less accurate.
All spread lines can be returned, not just the single maximum spread line.
There is an option to ensure that the spread line is completely contained within the second fire progression polygon, but this works very slowly and can not always find an appropriate solution
min_minutes and max_minutes are used to set that max and min timestep between polygons. When there are two polygons with a difference in time outside the range, the process is not run.
The vertices on the boundary of the polygons are converted to points to find the max spread line. densify_m how often extra vertices should be added to the polygon boundary, especially on staight boundary sections, before the polygon is converted to points.
Have a smaller value (metres) will produce more accurate results, but will take more time to run.
}
\examples{
#
}
