% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_max_spread_line.R
\name{fire_max_spread_line}
\alias{fire_max_spread_line}
\title{Find maximum fire spread line}
\usage{
fire_max_spread_line(
  polygon,
  time_col,
  include_spots = F,
  include_backburns = F
)
}
\arguments{
\item{polygon}{SF polygons. Polygon at time two should completely cover time 1 polygon. Time 3 completely cover time 2 etc.}

\item{time_col}{Name of date time column (posix)}

\item{include_spots}{T/F to include spot fire in spread line creation. Data must have firetype column populated with "spot", "main" or "backburn"}

\item{include_backburns}{T/F to include backburns in spread line creation. Data must have firetype column populated with "spot", "main" or "backburn"}
}
\value{
sf object
}
\description{
Find maximum fire spread line
}
\examples{
#
}
