% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_RANN_nearest_points.R
\name{fire_RANN_nearest_points}
\alias{fire_RANN_nearest_points}
\title{Fire nearest neighbour lines using RANN packages}
\usage{
fire_RANN_nearest_points(poly, poly.prior, densify_m, max_only = T)
}
\arguments{
\item{poly}{sf polygon of fire progression}

\item{poly.prior}{sf polygon of fire progression, the one captured before poly}

\item{densify_m}{minimum distance between vertices on the polygon for which to search for maximum spread line}

\item{max_only}{T/F return only single maximum spread line}
}
\value{
sf lines
}
\description{
Fire nearest neighbour lines using RANN packages
}
\examples{
#
}
