% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_sample_agcd.R
\name{fire_sample_agcd}
\alias{fire_sample_agcd}
\title{Sample monthly precipitation from AGCD}
\usage{
fire_sample_agcd(dat, time_utc_col, buffer_m)
}
\arguments{
\item{dat}{sf object with geometry and time column}

\item{time_utc_col}{character of time column name}
}
\value{
data frame
}
\description{
Sample monthly precipitation from AGCD
}
\examples{
#
}
