% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_search_landsat.R
\name{fire_search_landsat}
\alias{fire_search_landsat}
\title{Search NCI thredds server for Landsat data}
\usage{
fire_search_landsat(
  fire_bbox,
  start_date,
  end_date,
  regex_filter = "odc-metadata.yaml|band07|band05|band04"
)
}
\arguments{
\item{fire_bbox}{bounding box sf polygon of a fire}

\item{start_date}{start date to search for landsat images (YYYY-mm-dd)}

\item{end_date}{last date to search for landsat images (YYYY-mm-dd)}
}
\value{
a data frame with download paths of landsat imagery
}
\description{
Search NCI thredds server for Landsat data
}
\examples{
#x <- fire_process_polygon(dat.fire.polygon,4283,"FireName","StartDate","EndDate",2,2,buffer_km = 20)
#z <- fire_search_landsat(x$fire_bbox,"2019-12-02","2019-12-10")
}
