% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_search_progs.R
\name{fire_search_progs}
\alias{fire_search_progs}
\title{Search for fire progression polygons}
\usage{
fire_search_progs(
  fire_bbox = wfprogression::fire_bbox_polygon(),
  start_time,
  end_time,
  return_geom = TRUE,
  dbpassword
)
}
\arguments{
\item{fire_bbox}{A polygon, usually fire bounding box, to search for images}

\item{start_time}{The first date for which to search for images YYYY-mm-dd}

\item{end_time}{The last date for which to search for images YYYY-mm-dd}

\item{return_geom}{TRUE return sf object, FALSE return data frame}
}
\value{
sf object
}
\description{
Search for fire progression polygons
}
\examples{
#dat.progs.sf <- fire_search_progs(fire_bbox=dat.bbox,start_time=datestart,end_time=dateend)
}
