% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_stac_download.R
\name{fire_stac_download}
\alias{fire_stac_download}
\title{Download Landsat and Sentinel 2 images.}
\usage{
fire_stac_download(fire_bbox, stac_df, dest_folder)
}
\arguments{
\item{fire_bbox}{sf polygon of search area, same polygon used as input to fire_stac_search()}

\item{stac_df}{Data frame that is the output of fire_search_stac}

\item{dest_folder}{the output folder for geotifs.}
}
\value{
Writes geotifs to file
}
\description{
Download landsat and sentinel 2  false colour images. Images are clipped to fire_bbox.
Requires that data frame output from fire_stac_search()
}
\examples{
#outdir <- "C:\\temp\\landsat_and_sentinel"
#dir.create(outdir)
#fire_download_stac(dat.bbox,dat.stac,outdir)
}
