% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_stac_search.R
\name{fire_stac_search}
\alias{fire_stac_search}
\title{Search for Landsat and Sentinel 2 data}
\usage{
fire_stac_search(
  fire_bbox,
  start_time,
  end_time,
  collection_names = c("ga_ls7e_ard_3", "ga_ls8c_ard_3", "ga_ls9c_ard_3",
    "ga_ls5t_ard_3", "ga_s2am_ard_3", "ga_s2bm_ard_3")
)
}
\arguments{
\item{fire_bbox}{A polygon, usually fire bounding box, to search for images}

\item{start_time}{The first time for which to search for images (posixct required). This will be converted to utc.}

\item{end_time}{The last time for which to search for images (posixct required). This will be converted to utc.}

\item{collection_names}{Names of which Landsat and Sentinel 2 products to search for. Defaults to all products from Landsat 5 onwards.}
}
\value{
data frame with image paths
}
\description{
Search Geoscience Aus. STAC browser for Landsat and Sentinel 2 images that intersect polygon and match search dates.
This will return a data frame of images path to input to fire_stac_download() function.
A max of 999 records will be returned, so location and or date may need to be limited.
Info on STAC browser is here: https://knowledge.dea.ga.gov.au/guides/setup/gis/stac/
}
\examples{
#dat.stac <- fire_search_stac(fire_bbox = dat.bbox,start_time = datestart,end_time = dateend)
}
