% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire_thredds_list.R
\name{fire_thredds_list}
\alias{fire_thredds_list}
\title{Internal function the list files in thredds https path}
\usage{
fire_thredds_list(catalog_url, regex)
}
\arguments{
\item{catalog_url}{catalog.html path}

\item{regex}{regex filter string}
}
\value{
file names
}
\description{
Internal function the list files in thredds https path
}
\examples{
 #dplyr::mutate(df,
 #file_name=purrr::map(path_catalog,~fire_thredds_list(.x,"\\.tif$|\\.yaml$|\\.jpg$|\\.sha1$|\\.json$")))
}
